refactor(db): split schema and migrations into focused modules
Extract SCHEMA/pragmas and migration logic from the 998-line db.ts
singleton so schema changes and migrations can be edited and tested
in isolation without touching the init wrapper.
